fastapi_server/requirements.txt
fastapi
uvicorn
httpx
motor
pydantic
tortoise-orm
asyncio
python-dotenv
asyncpg